requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
attrs==25.4.0
bcrypt==5.0.0
bitarray==3.8.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
ckzg==2.1.7
click==8.3.1
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.3.1
hexbytes==1.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.11
MarkupSafe==3.0.3
marshmallow==4.3.0
multidict==6.7.1
packaging==26.0
parsimonious==0.10.0
propcache==0.4.1
py-solc-x==2.0.5
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
pyunormalize==17.0.0
regex==2026.2.28
requests==2.32.5
rlp==4.1.0
SQLAlchemy==2.0.46
toolz==1.1.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
web3==7.14.1
websockets==15.0.1
Werkzeug==3.1.5
yarl==1.23.0
